# SailSight.Api/OpenApi/SailSight.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GetDocument.Insider | v1",
    "description": "GetDocument Command-line Tool inside man",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/Marks/{id}": {
      "get": {
        "tags": [
          "Marks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MarkDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Marks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMarkRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMarkRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MarkDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Marks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MarkDto": {
        "required": [
          "id",
          "name",
          "activeFrom",
          "activeUntil",
          "latitude",
          "longitude",
          "description"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "activeFrom": {
            "type": "string",
            "format": "date"
          },
          "activeUntil": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateMarkRequest": {
        "required": [
          "name",
          "activeFrom",
          "activeUntil",
          "latitude",
          "longitude",
          "description"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "activeFrom": {
            "type": "string",
            "format": "date"
          },
          "activeUntil": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}